+++ src/cosserat-rod.cc
@@ -220,7 +220,7 @@ int main (int argc, char *argv[]) try
   //////////////////////////////////////////////
 
   using ATargetSpace = TargetSpace::rebind<adouble>::other;
-  using GeodesicInterpolationRule  = GFE::LocalGeodesicFEFunction<1, double, ScalarBasis::LocalView::Tree::FiniteElement, ATargetSpace>;
+  using GeodesicInterpolationRule  = GFE::LocalGeodesicFEFunction<ScalarBasis, ATargetSpace>;
   using ProjectedInterpolationRule = GFE::LocalProjectedFEFunction<1, double, ScalarBasis::LocalView::Tree::FiniteElement, ATargetSpace>;
 
   // Assembler using ADOL-C
@@ -313,7 +313,7 @@ int main (int argc, char *argv[]) try
   for (size_t i=0; i<x.size(); ++i)
     orientationConfiguration[i] = x[i][_1];
 
-  using RotationInterpolationRule = GFE::LocalGeodesicFEFunction<1, double, ScalarBasis::LocalView::Tree::FiniteElement, GFE::Rotation<double,3> >;
+  using RotationInterpolationRule = GFE::LocalGeodesicFEFunction<ScalarBasis, GFE::Rotation<double,3> >;
 
   GFE::EmbeddedGlobalGFEFunction<ScalarBasis, RotationInterpolationRule,GFE::Rotation<double,3> > orientationFunction(scalarBasis,
                                                                                                                       orientationConfiguration);