+++ src/harmonicmaps.cc
@@ -288,7 +288,7 @@ int main (int argc, char *argv[])
     DUNE_THROW(Exception, "Unknown energy type '" << energy << "'");
 
   // Next: The local energy, i.e., the integral of the density over one element
-  using GeodesicInterpolationRule  = GFE::LocalGeodesicFEFunction<dim, double, FEBasis::LocalView::Tree::FiniteElement, ATargetSpace>;
+  using GeodesicInterpolationRule  = GFE::LocalGeodesicFEFunction<FEBasis, ATargetSpace>;
   using ProjectedInterpolationRule = GFE::LocalProjectedFEFunction<dim, double, FEBasis::LocalView::Tree::FiniteElement, ATargetSpace>;
 
   std::shared_ptr<GFE::LocalEnergy<FEBasis,ATargetSpace> > localEnergy;